Fix component positioning in KiCAD-nightly
Component positions were way off in KiCAD-nightly.

Opening converted schematics in KiCAD 4.0.2, saving them there and then
opening the scematics in KiCAD-nightly worked.

One of the differences shown in a 'diff' was the addition of these "P"
tags by 4.0.2 and this simple change seems to have fixed the schematics
for the Cryptech Alpha schematics at least.