Properly put values on components.
Record type '41', name 'Value' used to produce a text label with the
value next to components. With this change the value is properly set on
the component, and the text label is skipped in favor of the usual
displaying of values on components.

The change was a bit invasive since I didn't want to repeat the whole
'find-x-y' code a third time, but rather refactored it into a subroutine
used from all three places that produce these 'F' lines.